requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
pypdf2
python-dotenv
requests
sentence-transformers
transformers
torch
langchain-huggingface 